tdb v2.1 unstable (API doc encoder and exstim)

+++ Encoder/encoder.cpp
@@ -9,9 +9,8 @@ void Encoder::updateValues(void)
     countPrev = count;
     count -= encode->getPulses();
     distance = count * WHEEL_DIAMETER * PI / CPR;
-    // distance += static_cast<float>(encode->getRevolutions()*WHEEL_DIAMETER*PI);
     countBuffer = count - countPrev; // difference in count
-    speed = ((static_cast<float>(countBuffer) / CPR) * SYS_OUTPUT_RATE) * WHEEL_DIAMETER * PI;
+    speed = ((static_cast<double>(countBuffer) / CPR) * SYS_OUTPUT_RATE) * WHEEL_DIAMETER * PI;
     encode->reset();
 };
 
@@ -28,6 +27,6 @@ void Encoder::resetDistance(void)
     distance = count = 0.0f;
 };
 
-float Encoder::getDist(void) { return distance; }; // returns distance from last reset() call
-float Encoder::getSpeed(void) { return speed; };   // returns speed in m/s
+double Encoder::getDist(void) { return distance; }; // returns distance from last reset() call
+double Encoder::getSpeed(void) { return speed; };   // returns speed in m/s
 int Encoder::getCount(void) { return count; };
